Fix remaining param sync issues and update param_defaults.py
- param_defaults.py: val_split_mode "random" -> "percent"
- routes.py: blocks_to_swap Form(22) -> Form(GENERATION_DEFAULTS[...]) (was 20/22 mismatch)
- TrainingConfig.tsx: vae_dtype fallback "fp32" -> "fp16" to match backend default
- TaggerTrainingConfig.tsx: val_split_mode default "random" -> "percent" for UI consistency
- TaggerTrainingConfig.tsx: fix TS error - dataset_configs cast via unknown
- TaggerTrainingConfig.tsx: fix TS error - loss_function ?? "" for .includes() call

+++ frontend/src/components/training/TrainingConfig.tsx
@@ -409,7 +409,7 @@ export default function TrainingConfig({ onClose, onRunCreated, editRunId, onRun
   const weightDtype = params.weight_dtype ?? "fp32";
   const trainingDtype = params.training_dtype ?? "fp16";
   const outputDtype = params.output_dtype ?? "fp32";
-  const vaeDtype = params.vae_dtype ?? "fp32";
+  const vaeDtype = params.vae_dtype ?? "fp16";
   const mixedPrecision = params.mixed_precision ?? true;
   const useFlashAttention = params.use_flash_attention ?? false;
   const minSnrGamma = params.min_snr_gamma ?? 5.0;
